Split out Pybind11 to allow using ILogger in Catch2-tests
